Removed the static tag in favor of a template.
It turns out that {% autocomplete_light_static %} worked great in the
test projects because luckily the channels were registered in the right
order.

Adding dependency management to the static file registry would just be
insane. Instead of bloating this feature, it was totally removed. Still,
to kickstart autocomplete_light in a project, the user can include
autocomplete_light/static.html.

Later on, the user may override this template or just load the
javascript dependencies directly.