Remove limitation of choices_for_values
Enforcing a limitation on the choices available to the already
selected values also enforces a limit on the amount of values
which can be selected. For example a limit_choices value of 20
would result in an error if more than 20 values were selected
using the autocompletion feature. Removing the limit also
removes this corner-case problem.