Added tests to detect bug described in GH269
A bug is triggered when a model is registered with autocomplete light
and widget_attrs are provided.  In this situation, the widget_attrs
dictionary was being passed back through the HTML rendering chain and
modified in-place (e.g. the TextInput for Django forms would add a
'value' attribute).  The 'value' attribute would then show up in the
original widget_attrs dictionary.  This test case ensures that a copy of
the widget_attrs dictionary is made so that the original dictionary is
not modified.