WIP Fix #608: V2-style on-the-fly object creation
It didn't work very well to create the object in the field's clean()
method. Another option was to do such thing in the widget's
value_from_datadict() method but there is no good way to know if a value
is meant to be an existing id or a new value anyway (ie. string ids).

This patch rewrites this feature in the same way it was implemented in
v2.

It breaks backward compatibility as it removes a few form field classes
in favor of a new argument in the view creation.

A new test app was added, select2_many_to_many_create, to test that
feature on multiple selects.

WIP: add tests and docs before merging.