Fix #631: Initializing GFKs with values from forms
Form field may be initialized with either a value from a POST or GET
request, either with a value returned by model field's
`value_from_object()` method.

Without this patch, initializing from a POST or GET request would fail
as reported in #631::

    'unicode' object has no attribute '_meta'

This is how Django's ModelChoiceField deals with that::

    def prepare_value(self, value):
        if hasattr(value, '_meta'):
            if self.to_field_name:
                return value.serializable_value(self.to_field_name)
            else:
                return value.pk

        # If it's not a model then it's a string from a POST / GET dict
        return super(ModelChoiceField, self).prepare_value(value)

This patch does the same for our GFK field.