Finish initial Select2List implementation.
Add two fields that can be used to avoid easy mistakes when using
Select2List with a ChoiceField since we do not support differences
between text and value choices.

Provide test coverage.

Fix existing formatting errors in docs tutorial sections "Autocompleting
based on a List of Strings" and update. Remove "Specifying Placeholder
Labels" as this is covered in "Passing options to select2".