Move create_option functionality into it's own method
Main motivation was wanting to override this behaviour without touching
any other parts of render_to_response but I think this is also a clearer
way to organise the code.

Actually - in most cases all I really want to change is the wording of
the create text. How would you feel about a get_create_label(self,
queryset) method?

(example use case - I if someone enters a string with an @ in it I
create a new contact with an email address, otherwise I assume it's a
name - I'd like to make this behaviour explicit via the create label)