dal: prefer pristine jQuery to the django one
First of all i can't see how select2 may work without a jQuery
as a global $. Said that now when i have another copy of jQuery
on top of the django one, select2 will register to $ but dal
will set yl.jQuery to the django one instead.
If we prefer using $ instead they'll both use the same jQuery
instance and magically work.
If there's not jQuery loaded as $ the code will behave as before.
This makes dal working reliably in the admin for me.