Fix #1129: documentation
This fixes the current warnings and adds a CI job to prevent such
warnings to occur again.

Also disables the travis job because it's not working anymore, despite
that it's still working great on my laptop which is enough for me, given
a pretty low contribution activity on this repo.