Add documentation on passing URL parameters
Hey,

Something I discovered very recently while working with this package was that it is indeed possible to pass parameters to the Autocomplete View through URL parameters in the widget of a form field. I've created a category on this in the tutorial as some people might find this use case helpful. It was very useful for myself and my company!

Thank you for an awesome tool,
Robert