Two isolated unadorned simple test cases added
Something sorely lacking from the test_project seems to be a really stripped down, isolated test of the single and multiselect widgets. Very useful and important I believe, for seeing the very bare minimum configuration for a working widget without a lot of very distracting dressing in templates and views. These a totally drab new pages to demonstrated the stripped down bare minimalist widgets in action.

That said, something is wrong, and this first cut serves as little more than a minimal stripped won demonstration of the focus bug being discussed on github now . It is either a DAL bug or an incorrect use of DAL, but this tiny little view will help anyone assessing it to offer commentary on what has gone wrong. There's almost nothing there and hence it is, as per the goal, free of a lot of distracting context.