Update taggit.rst docs.
For `taggit` the create option is broken, and it is unnecessary.

One of the drop down options includes the new tag. So if you've typed `funny` and `funny` does not exist you'll have:

```
funny
Create 'funny'
```

Picking `funny` works as intended -- it creates a new tag properly.

Picking `Create 'funny'` creates two tags and labels them wonkily.

This change removes the create option entirely from the docs so this works out of the box for end-users.